Revert "feat(async-modules): added asynchronous calls + spliting modules"
