--- src/helpers/common.ts
@@ -1,4 +1,0 @@
-export const getCurrentYear = () => {
-    const d = new Date();
-    return d.getFullYear();
-};
