+++ src/helpers/handlebars.ts
@@ -5,6 +5,13 @@ export const isLink = (exp: string, got: string): boolean => {
 export const ifEquals = (a: string, b: string): boolean => {
   return a === b;
 };
+export const displayPagesNumber = (interval: number,current: number, pages: number ): Array<number> => {
+  let result = [];
+  for (; interval <= (current+ 4) && interval <= pages; interval++){
+    result.push(interval);
+  }
+  return result;
+};
 
 export const isLower = (v1: number,v2: number): boolean => {
   return v1 < v2;
@@ -18,6 +25,15 @@ export const ifDifferent = (a: number, b: number): boolean => {
   return a !== b;
 };
 
+export const constructUrl = (url: string, param: string): string => {
+  let data = url.includes("&") ? url.split("&") : url.split("?");
+  for(let i = 0; i< data.length;i++){
+    if(data[i].includes("page")){
+      data[i] = "page="+param;
+    }
+  }
+  return url.includes("&") ? data.join("&") : data.join("?");
+}
 export const contains = (a: string, b:string): boolean => {
   return a.includes(b);
 }
