+++ src/views/layouts/main.hbs
@@ -19,7 +19,6 @@
 
       gtag('config', 'G-WB7BVLNTNP');
     </script>
-
     <style>
         .dev-card {
             border: solid 1px rgba(51, 51, 51, 0.25);
