--- src/views/partials/app-list.hbs
@@ -1,5 +1,0 @@
-<div class="container mt-5" style="margin-bottom: 25px">
-    <div class="row g-5" id="app-list"></div>
-</div>
-<!-- The js part -->
-<script src="/scripts/app-list.js"></script>
