+++ src/views/partials/home.hbs
@@ -1,0 +1,116 @@
+<form class="row pt-4 pb-4 g-2 d-flex justify-content-center mt-5 mb-5 bg-light">
+    <div class="col-md-3">
+	  <div class="input-group">
+    	<input id="search-input" type="text" id="query" name="query" class="form-control" placeholder="search" value="{{query}}"/>
+  	    <button type="submit" class="btn btn-primary">
+  	  	  <i class="fa fa-search" aria-hidden="true"></i>
+  	    </button>
+	  </div>
+	</div>
+
+	<div class="col-md-2">
+      <select class="form-select" id="category" name="category" aria-label="Default select category">
+        {{#each categories.list}}
+          {{#if (ifEquals this @root.categories.selected) }}
+            <option value="{{ this }}" selected>{{this}}</option>
+          {{else}}
+            <option value="{{ this }}">{{this}}</option>
+          {{/if}}
+        {{/each}}
+      </select>
+    </div>
+
+	<div class="col-md-2">
+      <select class="form-select form-control" id="tag" name="tag" aria-label="Default select tag">
+        {{#each tags.list}}
+          {{#if (ifEquals this @root.tags.selected) }}
+            <option value="{{ this }}" selected>{{this}}</option>
+          {{else}}
+            <option value="{{ this }}">{{this}}</option>
+          {{/if}}
+        {{/each}}
+      </select>
+    </div>
+</form>
+
+<div class="container mt-5" style="margin-bottom: 25px">
+    <div class="row g-5">
+        {{#each apps}}
+            <div class="col-md-4 d-flex justify-content-center">
+              <div class="card h-100 w-100">
+
+                <div class="card-body">
+                  <div class="d-flex justify-content-start align-items-center mb-2">
+                    <img style="height: 50px; width: 50px" class="rounded" src="{{this.image}}" class="card-img-top" alt="image">
+                    <div class="ms-2 pt-3">
+                        <a style="text-decoration: underline; color: black" href="{{this.url}}" target="_blank">
+                          <h6 class="card-title"><strong>{{this.title}}</strong></h6>
+                        </a>
+                        <p style="font-size: 0.8rem"><em>{{this.category}}</em></p>
+                      </div>
+                  </div>
+                    <p style="max-height: 150px; white-space: wrap; overflow: hidden;" class="card-text">{{this.description}}</p>
+                </div>
+
+                <div class="row g-1 card-footer row-cols-auto d-flex align-items-center">
+                 {{# if this.website}}
+                    <a class="col" href="{{this.website}}" target="_blank">
+                      <i class="fa fa-globe fa-lg" aria-hidden="true"></i>
+                    </a>
+                  {{/if}}
+
+                  {{# if this.telegram}}
+                    <a class="col" href="{{this.telegram}}" target="_blank">
+                      <i class="fa fa-telegram fa-lg" aria-hidden="true"></i>
+                    </a>
+                  {{/if}}
+
+                  {{# if this.whatsapp}}
+                    <a class="col" href="{{this.whatsapp}}" target="_blank">
+                      <i class="fa fa-whatsapp fa-lg" aria-hidden="true"></i>
+                    </a>
+                  {{/if}}
+
+                  {{# if this.github_account}}
+                    <a class="col" href="{{this.github_account}}" target="_blank">
+                      <i class="fa fa-github-square fa-lg" aria-hidden="true"></i>
+                    </a>
+                  {{/if}}
+
+                  {{# if this.twitter}}
+                    <a class="col" href="{{this.twitter}}" target="_blank">
+                      <i class="fa fa-twitter-square fa-lg" aria-hidden="true"></i>
+                    </a>
+                  {{/if}}
+
+                  {{# if this.facebook}}
+                    <a class="col" href="{{this.facebook}}" target="_blank">
+                      <i class="fa fa-facebook-square fa-lg" aria-hidden="true"></i>
+                    </a>
+                  {{/if}}
+
+                  {{# if this.appstore}}
+                    <a class="col" href="{{this.appstore}}" target="_blank">
+                      <i class="fa fa-apple fa-lg" aria-hidden="true"></i>
+                    </a>
+                  {{/if}}
+
+                  {{# if this.playstore}}
+                    <a class="col" href="{{this.playstore}}" target="_blank">
+                      <i class="fa fa-android fa-lg" aria-hidden="true"></i>
+                    </a>
+                  {{/if}}
+
+                   {{# if this.slack}}
+                    <a class="col" href="{{this.slack}}" target="_blank">
+                      <i class="fa fa-slack fa-lg" aria-hidden="true"></i>
+                    </a>
+                  {{/if}}
+
+                </div>
+
+              </div>
+            </div>
+        {{/each}}
+    </div>
+</div>
